Keep InCallUI in memory (1/3)
In lower end devices, startup of the InCallUI process can take
up to 10 seconds when the device is busy, which prevents the user
from taking the call. Keep the process in memory like previous
versions where they are a part of Phone.apk.

This converts InCallUI into a dummy apk with the persistent
attribute set to true and allows the activity from Dialer.apk,
which uses the com.android.incallui process to be loaded into the
persistent process.

Change-Id: Ic6a0461bb0373cd21213a03fec41c98b1e65ee40
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>