Layers 4.1: InCallUI dialpad digits color
InCallUI's dialpad digits (not the digits' text) color is linked to the InCallUI theme primary color, on runtime. This is coded as such because Google wanted to update the InCallUI's theme (primary and primary-dark) colors in MSIM devices on the fly, based on which SIM card was used to make the call.
This commit:
* makes this behavior configurable by a boolean
* links the digits color to a color value

This way, the default behavior is preserved. The themer has to set the "config_dialpadDigitsStaticColor" boolean to true, in order to theme the color value. If he doesn't change the boolean, the default behavior is preserved and the digits color works as Google wanted.

Change-Id: I87aad13969a935791cf6ae82ee87652307fe493c